app/db.py
from pymongo import MongoClient
import os
from dotenv import load_dotenv

load_dotenv()

client = MongoClient(os.getenv("MONGO_URL"))
db = client["attendance_db"]

users = db["users"]
attendance = db["attendance"]