requirements.txt
fastapi
uvicorn
python-multipart
pymongo
python-dotenv
passlib[bcrypt]
numpy
opencv-contrib-python